update room setting page

+++ oop_project/src/Models/Cal_Watercost.java
@@ -2,15 +2,15 @@
 package Models;
 
 public class Cal_Watercost extends Calculator {
-    public static double Water_CostPerUnits = 25;
+    public static int Water_CostPerUnits = 25;
     
-    public Cal_Watercost(double meter) {
+    public Cal_Watercost(int meter) {
         super(Water_CostPerUnits, meter, 0);
         this.totalcost = CalculateCost();
     }
 
     @Override
-    public double CalculateCost() {
+    public int CalculateCost() {
         return Water_CostPerUnits * meter;
     }
