+++ oop_project/src/Models/Calculate.java
@@ -1,5 +1,5 @@
 package Models;
 
 public interface Calculate {
-    public double CalculateCost();
+    public int CalculateCost();
 }
